requirements.txt
pytest>=5.4.1
numpy==1.16.2
pandas==0.23.4
ipython==7.8.0
ipykernel==5.1.2
azureml-sdk>=1.1.5
scikit-learn==0.20.3